Refactor for L0C
