+++ disdrodb/cli/disdrodb_run_l0.py
@@ -20,11 +20,9 @@ from typing import Optional
 
 import click
 
-from disdrodb.l0.routines import (
-    click_l0_archive_options,
-)
 from disdrodb.utils.cli import (
     click_base_dir_option,
+    click_l0_archive_options,
     click_processing_options,
     click_stations_options,
     parse_arg_to_list,
@@ -47,7 +45,7 @@ def disdrodb_run_l0(
     # L0 archive options
     l0a_processing: bool = True,
     l0b_processing: bool = True,
-    l0b_concat: bool = True,
+    l0c_processing: bool = True,
     remove_l0a: bool = False,
     remove_l0b: bool = False,
     # Processing options
@@ -87,17 +85,14 @@ def disdrodb_run_l0(
     l0b_processing : bool
         Whether to launch processing to generate L0B netCDF4 file(s) from L0A data.
         The default is True.
-    l0b_concat : bool
-        Whether to concatenate all raw files into a single L0B netCDF file.
-        If l0b_concat=True, all raw files will be saved into a single L0B netCDF file.
-        If l0b_concat=False, each raw file will be converted into the corresponding L0B netCDF file.
-        The default is False.
+    l0c_processing : bool
+        Whether to launch processing to generate L0C netCDF4 file(s) from L0C data.
+        The default is True.
     remove_l0a : bool
         Whether to keep the L0A files after having generated the L0B netCDF products.
         The default is False.
     remove_l0b : bool
-         Whether to remove the L0B files after having concatenated all L0B netCDF files.
-         It takes places only if l0b_concat = True
+         Whether to remove the L0B files after having produced L0C netCDF files.
         The default is False.
     force : bool
         If True, overwrite existing data into destination directories.
@@ -123,7 +118,7 @@ def disdrodb_run_l0(
         Format: <...>/DISDRODB
         If not specified, uses path specified in the DISDRODB active configuration.
     """
-    from disdrodb.l0.routines import run_disdrodb_l0
+    from disdrodb.routines import run_disdrodb_l0
 
     # Parse data_sources, campaign_names and station arguments
     base_dir = parse_base_dir(base_dir)
@@ -140,7 +135,7 @@ def disdrodb_run_l0(
         # L0 archive options
         l0a_processing=l0a_processing,
         l0b_processing=l0b_processing,
-        l0b_concat=l0b_concat,
+        l0c_processing=l0c_processing,
         remove_l0a=remove_l0a,
         remove_l0b=remove_l0b,
         # Processing options
