+++ disdrodb/l0/l0c_processing.py
@@ -32,6 +32,24 @@ logger = logging.getLogger(__name__)
 
 
 def get_files_per_days(filepaths):
+    """
+    Organize files by the days they cover based on their start and end times.
+
+    Parameters
+    ----------
+    filepaths : list of str
+        List of file paths to be processed.
+
+    Returns
+    -------
+    dict
+        Dictionary where keys are days (as strings) and values are lists of file paths
+        that cover those days.
+
+    Notes
+    -----
+    This function adds a tolerance of 60 seconds to account for imprecise time logging by the sensors.
+    """
     # Retrieve file start_time and end_time
     files_start_time, files_end_time = get_start_end_time_from_filepaths(filepaths)
 
@@ -58,8 +76,6 @@ def get_files_per_days(filepaths):
     # Use broadcasting to create a boolean matrix indicating which files cover which days
     mask = (files_start_time <= days) & (files_end_time >= days)  # shape (n_files, n_days)
 
-    # np.sum(mask, axis=0)
-
     # Build a mapping from days to file indices
     # For each day (column), find the indices of files (rows) that cover that day
     dict_days = {}
@@ -73,6 +89,29 @@ def get_files_per_days(filepaths):
 
 
 def find_isel_common_time(da1, da2):
+    """
+    Find the indices of common time steps between two data arrays.
+
+    Parameters
+    ----------
+    da1 : xarray.DataArray
+        The first data array with a time coordinate.
+    da2 : xarray.DataArray
+        The second data array with a time coordinate.
+
+    Returns
+    -------
+    da1_isel : numpy.ndarray
+        Indices of the common time steps in the first data array.
+    da2_isel : numpy.ndarray
+        Indices of the common time steps in the second data array.
+
+    Notes
+    -----
+    This function assumes that both input data arrays have a "time" coordinate.
+    The function finds the intersection of the time steps in both data arrays
+    and returns the indices of these common time steps for each data array.
+    """
     intersecting_timesteps = np.intersect1d(da1["time"], da2["time"])
     da1_isel = np.where(np.isin(da1["time"], intersecting_timesteps))[0]
     da2_isel = np.where(np.isin(da2["time"], intersecting_timesteps))[0]
@@ -80,6 +119,27 @@ def find_isel_common_time(da1, da2):
 
 
 def check_same_raw_drop_number_values(list_ds, filepaths):
+    """
+    Check if the 'raw_drop_number' values are the same across multiple datasets.
+
+    This function compares the 'raw_drop_number' values of multiple datasets to ensure they are identical
+    at common timesteps.
+
+    If any discrepancies are found, a ValueError is raised indicating which files
+    have differing values.
+
+    Parameters
+    ----------
+    list_ds : list of xarray.Dataset
+        A list of xarray Datasets to be compared.
+    filepaths : list of str
+        A list of file paths corresponding to the datasets in `list_ds`.
+
+    Raises
+    ------
+    ValueError
+        If 'raw_drop_number' values differ at any common timestep between any two datasets.
+    """
     # Retrieve variable to compare
     list_drop_number = [ds["raw_drop_number"].compute() for ds in list_ds]
     # Compare values
@@ -96,6 +156,37 @@ def check_same_raw_drop_number_values(list_ds, filepaths):
 
 
 def create_daily_file(day, filepaths):
+    """
+    Create a daily file by merging and processing data from multiple filepaths.
+
+    Parameters
+    ----------
+    day : str or numpy.datetime64
+        The day for which the daily file is to be created.
+        Should be in a format that can be converted to numpy.datetime64.
+    filepaths : list of str
+        List of filepaths to the data files to be processed.
+
+    Returns
+    -------
+    xarray.Dataset
+        The processed dataset containing data for the specified day.
+
+    Raises
+    ------
+    ValueError
+        If less than 5 timesteps are available for the specified day.
+
+    Notes
+    -----
+    - The function adds a tolerance for searching timesteps
+    before and after 00:00 to account for imprecise logging times.
+    - It checks that duplicated timesteps have the same raw drop number values.
+    - The function infers the time integration sample interval and
+    regularizes timesteps to handle trailing seconds.
+    - The data is loaded into memory and connections to source files
+    are closed before returning the dataset.
+    """
     # Define start day and end of day
     start_day = np.array(day).astype("M8[D]")
     end_day = start_day + np.array(1, dtype="m8[D]") - np.array(1, dtype="m8[s]")  # avoid 00:00 of next day !
@@ -106,7 +197,8 @@ def create_daily_file(day, filepaths):
     end_day_tol = end_day + np.array(60, dtype="m8[s]")
 
     # Open files with data within the provided day
-    # list_ds = [xr.open_dataset(filepath, chunks={}).sel({"time": slice(start_day_tol, end_day_tol)}) for filepath in filepaths]
+    # list_ds = [xr.open_dataset(filepath, chunks={}).sel({"time": slice(start_day_tol, end_day_tol)})
+    # for filepath in filepaths]
     list_ds = [xr.open_dataset(filepath, chunks={}).sortby("time") for filepath in filepaths]
     list_ds = [ds.sel({"time": slice(start_day_tol, end_day_tol)}) for ds in list_ds]
